refactored model out of the main file, renamed the main file
