+++ .gitignore
@@ -1,0 +1,5 @@
+.DS_Store
+.history
+.bak
+.idea
+.vscode/
