+++ model/document_form_field.go
@@ -1,0 +1,13 @@
+package model
+
+type DocumentFormField struct {
+	APIId    string `json:"api_id"`
+	Name     string `json:"name"`
+	Type     string `json:"type"`
+	X        int    `json:"x"`
+	Y        int    `json:"y"`
+	Width    int    `json:"width"`
+	Height   int    `json:"height"`
+	Required bool   `json:"required"`
+	Signer   int    `json:"signer"`
+}
