+++ model/embedded_signature_response.go
@@ -1,0 +1,5 @@
+package model
+
+type EmbeddedSignatureResponse struct {
+	Embedded *SignURLResponse `json:"embedded"`
+}
