+++ model/list_response.go
@@ -1,0 +1,6 @@
+package model
+
+type ListResponse struct {
+	ListInfo          *ListInfo           `json:"list_info"`
+	SignatureRequests []*SignatureRequest `json:"signature_requests"`
+}
