+++ model/signature_request_response.go
@@ -1,0 +1,5 @@
+package model
+
+type SignatureRequestResponse struct {
+	SignatureRequest *SignatureRequest `json:"signature_request"`
+}
