+++ model/warning.go
@@ -1,0 +1,6 @@
+package model
+
+type Warning struct {
+	Message string `json:"warning_msg"`
+	Name    string `json:"warning_name"`
+}
